packages/api-tools/requirements.txt
Flask
Flask-Cors
gunicorn
python-dotenv
firebase-admin
google-cloud-firestore
google-generativeai
beautifulsoup4
selenium
webdriver-manager